Move CpuId and CpuBase to common crate for arch independence
Moves CpuId (from arch) and CpuBase with offset constants (from sys)
to the common crate. This breaks the circular dependency between sys
and arch, enabling assembly files to be moved to arch in a follow-up
while preserving compile-time offset_of!() safety.

arch and sys re-export these types for backward compatibility.
